+++ platform/src/lib.rs
@@ -5,4 +5,12 @@ impl Platform {
     pub fn new() -> Self {
         Self
     }
+
+    #[cfg(target_os = "windows")]
+    pub fn start_keyboard_hook(&self) -> anyhow::Result<windows::KeyboardHook> {
+        windows::start_keyboard_hook()
+    }
 }
+
+#[cfg(target_os = "windows")]
+pub mod windows;