+++ shared_types/src/lib.rs
@@ -2,4 +2,4 @@ pub mod config;
 pub mod events;
 
 pub use config::Config;
-pub use events::AppEvent;
+pub use events::{AppEvent, KeyboardEvent};